requirements.txt
aiogram==3.3.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
gspread-asyncio==1.1.0
google-auth==2.26.1
pydantic-settings==2.1.0
apscheduler==3.10.4
loguru==0.7.2
gspread-formatting==1.2.0